[FIX] Add getField function to jwt.go
